requirements.txt
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.5
django-bootstrap-pagination==1.7.1
django-filter==2.2.0
django-heroku==0.3.1
django-robokassa-payments==1.3.3
get==2019.4.13
gunicorn==20.0.4
idna==2.8
lxml==4.4.2
numpy==1.18.0
pandas==0.25.3
Pillow==6.1.0
post==2019.4.13
psycopg2==2.8.3
public==2019.4.13
python-dateutil==2.8.1
python-dotenv==0.10.5
pytz==2019.2
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
whitenoise==5.0.1
